+++ kwcoco/exceptions.py
@@ -17,3 +17,11 @@ class InvalidAddError(ValueError):
     Error when trying to invalid data
     """
     pass
+
+
+class CoordinateCompatibilityError(ValueError):
+    """
+    Error when trying to perform operations on data in different coordinate
+    systems.
+    """
+    pass